requirements.txt
astroid==2.4.2
autopep8==1.5.3
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.3
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
